# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Finboard API
  version: 1.0.0
  description: |
    Finboard KYC, banking, and investment platform API.

    **Base URL:** http://localhost:4000 (API Gateway)

    **Authorize protected routes:** Click Authorize, enter `Bearer <token>` from signin or verify-otp.

    ---

    ## Local dev OTP

    A **random 6-digit OTP** is generated each time. In local dev, check the auth-service console log (`[DEV] Phone OTP for ...`). OTP is **never** returned in API responses.

    Run `pnpm seed:admin` before testing seeded accounts.

    ---

    ## Flow A — New user registration (test in order)

    | Step | Endpoint | Body highlights |
    |------|----------|-----------------|
    | 1 | POST /api/auth/signup | name, email, phone, password — **no OTP** |
    | 2 | POST /api/auth/verify-otp | same phone + otp `123456` → **JWT** |
    | — | POST /api/auth/send-otp | optional — resend OTP if expired |

    Use a **new** email/phone each signup test (e.g. newuser@finboard.local, +919900007777).

    ---

    ## Flow B — Email login (seeded or verified user)

    | Step | Endpoint | Values |
    |------|----------|--------|
    | 1 | POST /api/auth/signin | user@finboard.local / User@12345 |

    ---

    ## Flow C — Phone login

    | Step | Endpoint | Values |
    |------|----------|--------|
    | 1 | POST /api/auth/send-otp | phone +919876543210 |
    | 2 | POST /api/auth/phone-login | phone + otp `123456` |

    ---

    ## Flow D — Admin login

...

    ---

...
servers:
...
security: []
paths:
...
  /api/kyc/me:
    get:
      tags: [KYC]
      summary: Get my KYC application
      operationId: getMyKyc
      security:
        - bearerAuth: []
      responses:
        "200":
          description: KYC application (or null)
          content:
            application/json:
              schema:
                type: object
                properties:
                  application:
                    $ref: "#/components/schemas/KycApplication"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/kyc/submit:
    post:
      tags: [KYC]
      summary: Submit KYC application
      description: |
        Multipart upload with PAN and Aadhaar document images.
        Name, PAN, and Aadhaar must match a seeded identity (`pnpm seed:kyc`).
      operationId: submitKyc
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [name, panNumber, aadhaarNumber, pan, aadhaar]
              properties:
                name:
                  type: string
                  example: Rahul Sharma
                panNumber:
                  type: string
                  example: ABCPS1234F
                aadhaarNumber:
                  type: string
                  example: "111222333445"
                pan:
                  type: string
                  format: binary
                aadhaar:
                  type: string
                  format: binary
      responses:
        "201":
          description: KYC submitted
          content:
            application/json:
              schema:
                type: object
                properties:
                  application:
                    $ref: "#/components/schemas/KycApplication"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /api/kyc/admin/applications/{id}/approve:
    post:
      tags: [KYC]
      summary: Approve KYC application
      operationId: approveKyc
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                remarks:
                  type: string
                  maxLength: 500
      responses:
        "200":
          description: Application approved
        "404":
          $ref: "#/components/responses/NotFound"
  /api/kyc/admin/applications/{id}/reject:
    post:
      tags: [KYC]
      summary: Reject KYC application
      operationId: rejectKyc
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [remarks]
              properties:
                remarks:
                  type: string
                  minLength: 10
                  maxLength: 500
      responses:
        "200":
          description: Application rejected
        "400":
          $ref: "#/components/responses/BadRequest"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT from any successful login:
        - Flow A: `POST /api/auth/verify-otp` (after signup)
        - Flow B: `POST /api/auth/signin`
        - Flow C: `POST /api/auth/phone-login`
        - Flow D: `POST /api/auth/admin/signin`

        In Scalar, click **Authorize** and enter: `Bearer <your-token>`
  parameters:
    ResourceId:
      name: id
      in: path
      required: true
      schema:
        type: string
  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Missing or invalid JWT
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      properties:
        message:
          type: string
...
    KycApplication:
      type: object
      properties:
        _id:
          type: string
        userId:
          type: string
        name:
          type: string
        panNumber:
          type: string
        aadhaarNumber:
          type: string
        status:
          type: string
          enum:
            - draft
            - pending_admin_review
            - approved
            - rejected
            - failed
        aiVerification:
          $ref: "#/components/schemas/AiVerification"
        adminRemarks:
          type: string
        createdAt:
          type: string
          format: date-time
    AiVerification:
      type: object
      properties:
        overallScore:
          type: integer
          minimum: 0
          maximum: 100
        recommendation:
          type: string
          enum: [approve, review, reject]
        fields:
          type: object
          additionalProperties: true
        alignments:
          type: object
          properties:
            userInputVsIdentity:
              type: integer
            ocrVsIdentity:
              type: integer
            imagesVsIdentity:
              type: integer
              nullable: true
        summary:
          type: string
        verificationSource:
          type: string
          example: mistral_vision
        model:
          type: string
        verifiedAt:
          type: string
          format: date-time
...